Adds tests for RuleFindingIntersectionAutomaton.
RuleFindingIntersectionAutomaton needed JUnit tests to ensure that it performs correctly. These are added. Includes test for all the functions that needed to be implemented (or marked unsupported) in order to fulfill the TreeAutomaton subclass requirements. Also includes a simple test for performance, by running the whole thing for the decomposition automaton of a 25 word sentence.