Adds TreeAutomaton that learns from variable indicating trees.
In order to deal with the learning of rule trees, we will have to actually keep track of their weight. The TreeAddingAutomaton is fed with rule trees and assigns weights to rule trees according to how often subtrees have been seen. In order to do that it creates intermediate states that keep exact track of the trees which are only linked to a start state, when something deemed a variable is encountered.