Completes specification of StateWeighter.
The StateWeighter class is now fully specified and will be usable for implementation by extending classes. In order to achieve this is was necessary to write code for sampling rule trees from states. Return values of the Tree sampling process are now rule trees in order to fit with the update protocol.