Adds sampling class for Malign Automata.
We will often need a sampler that can deal with automata that have dead-ends and which are much to big to compute any inside scores.The new class gives a sampler that handles these two problems by creating an auxiliary benign automaton. It creates this by sampling top down and 'going back' on choices that lead nowhere. It still will not help with automata that allow for improper weighting.