Adds tests for Propagator.
The new, streamlined version of Propagator needed a new set of tests, which have been supplied now. The tests lead to the
removal of a number of bugs in the way that the generated automaton and alignments are constructed. Usually a state from the
original automaton was used where one from the constructed one was needed or vice versa.