Complete construction for Kuhlmannalgebra decomposition automaton and correct handling of ncg under connection.
The problem with the connection was that the last node of the first graph and the first node of the last graph where not merged.
Correcting this leads to the problem of constructing length 2 graphs. This has been handled by doing no merging iff the first graph
has only one (non-root) node. Also, happy new Year.