updated IsiAMR parser to work properly if reentrancy is before the original node in the string
This is a bit of a borderline case where it's not sure what the intended behavior should be, but this change will ensure compatibility with the AMR strings generated by github.com/ablodge/amr-utils. The case where it fails is if an anonymous node has a label that looks like a nodename (matching [a-z][0-9]*). To my understanding though, all such labels would be in quotation marks in the IsiAMR format, avoiding the issue.

I also added a couple more tests, and changed a test to reflect the new behavior. Also, ":polite +" markers in AMRs can now be properly read.